fix: fix storybook-config and move directory
