+++ app/components/assets/figma-plugin.png
